+++ include/visionaray/texture/texture.h
@@ -56,85 +56,10 @@ inline auto tex3D(Tex const& tex, vector<3, FloatT> const& coord)
 }
 
 
-#ifdef __CUDACC__
-
-template <
-    typename T,
-    cudaTextureReadMode ReadMode = cudaTextureReadMode(detail::tex_read_mode_from_type<T>::value)
-    >
-VSNRAY_GPU_FUNC
-inline typename cuda::map_texel_type<typename cuda_texture_ref<T, 1>::cuda_type, ReadMode>::vsnray_return_type
-tex1D(cuda_texture_ref<T, 1> const& tex, float coord)
-{
-    using tex_type          = cuda_texture_ref<T, 1>;
-    using cuda_type         = typename tex_type::cuda_type;
-    using return_type       = typename cuda::map_texel_type<cuda_type, ReadMode>::vsnray_return_type;
-    using cuda_return_type  = typename cuda::map_texel_type<cuda_type, ReadMode>::cuda_return_type;
-
-    cuda_return_type retval;
-
-    ::tex1D(
-            &retval,
-            tex.texture_object(),
-            coord
-            );
-
-    return cuda::cast<return_type>(retval);
-}
-
-template <
-    typename T,
-    cudaTextureReadMode ReadMode = cudaTextureReadMode(detail::tex_read_mode_from_type<T>::value)
-    >
-VSNRAY_GPU_FUNC
-inline typename cuda::map_texel_type<typename cuda_texture_ref<T, 2>::cuda_type, ReadMode>::vsnray_return_type
-tex2D(cuda_texture_ref<T, 2> const& tex, vector<2, float> coord)
-{
-    using tex_type          = cuda_texture_ref<T, 2>;
-    using cuda_type         = typename tex_type::cuda_type;
-    using return_type       = typename cuda::map_texel_type<cuda_type, ReadMode>::vsnray_return_type;
-    using cuda_return_type  = typename cuda::map_texel_type<cuda_type, ReadMode>::cuda_return_type;
-
-    cuda_return_type retval;
-
-    ::tex2D(
-            &retval,
-            tex.texture_object(),
-            coord.x,
-            coord.y
-            );
-
-    return cuda::cast<return_type>(retval);
-}
-
-template <
-    typename T,
-    cudaTextureReadMode ReadMode = cudaTextureReadMode(detail::tex_read_mode_from_type<T>::value)
-    >
-VSNRAY_GPU_FUNC
-inline typename cuda::map_texel_type<typename cuda_texture_ref<T, 3>::cuda_type, ReadMode>::vsnray_return_type
-tex3D(cuda_texture_ref<T, 3> const& tex, vector<3, float> coord)
-{
-    using tex_type          = cuda_texture_ref<T, 3>;
-    using cuda_type         = typename tex_type::cuda_type;
-    using return_type       = typename cuda::map_texel_type<cuda_type, ReadMode>::vsnray_return_type;
-    using cuda_return_type  = typename cuda::map_texel_type<cuda_type, ReadMode>::cuda_return_type;
-
-    cuda_return_type retval;
-
-    ::tex3D(
-            &retval,
-            tex.texture_object(),
-            coord.x,
-            coord.y,
-            coord.z
-            );
-
-    return cuda::cast<return_type>(retval);
-}
-
-#endif // __CUDACC__
-
 } // visionaray
+
+#ifdef __CUDACC__
+#include "cuda_texture.h"
+#endif // __CUDACC__
 
 #endif // VSNRAY_TEXTURE_H